+++ src/types/dummy/gi/Adw.d.ts
@@ -1,9 +1,12 @@
 declare namespace Adw {
     type ActionRow = any;
     type ComboRow = any;
+    type HeaderBar = any;
     type PreferencesGroup = any;
     type PreferencesPage = any;
     type PreferencesWindow = any;
+    type Toast = any;
+    type ToastOverlay = any;
 }
 
 declare const Adw: any;