+++ src/types/dummy/gi/Gtk.d.ts
@@ -1,9 +1,12 @@
 declare namespace Gtk {
+    type Button = any;
     type EventControllerKey = any;
     type Dialog = any;
+    type Label = any;
     type ShortcutController = any;
     type ShortcutManager = any;
     type ToggleButton = any;
+    type Widget = any;
 }
 
 declare const Gtk: any;